Switch to Alpine for the Docker image
If possible it is preferable to use Alpine for our Docker images as its
design is well-suited for Docker images. Since this is a straight-
forward image that does not rely on packages with compiled elements we
can safely switch from Debian to Alpine for the OS of the underlying
Docker image.