Store package versions as Docker environment variables
Update the Docker configuration to use environment variables to house
the versions of pip, pipenv, setuptools, and wheel that are installed
directly in the Dockerfile. This will DRY out the configuration a bit
as well as ensuring that the same versions are used for packages that
are installed multiple times.

The environment variable name format is cribbed from the official
Python Docker image configuration:
https://github.com/docker-library/python/blob/0a56f04a07495e0ed61663c73bc0c2e175773ddc/3.11/alpine3.17/Dockerfile#L132-L135
I tested without the leading `PYTHON_` in the names and only observed a
failure for usage of `PIPENV_VERSION`. Following the above mentioned
standard seemed like a reasonable approach as a result.

Co-authored-by: Shane Frasier <jeremy.frasier@trio.dhs.gov>